updated cleanuptask

+++ src/main/java/de/hhu/bsinfo/dxram/ms/tasks/mergesortapplication/CleanUpTask.java
@@ -13,21 +13,20 @@ import java.io.IOException;
 public class CleanUpTask implements Task {
     @Override
     public int execute(TaskContext p_ctx) throws IOException {
-        ChunkService chunkService = p_ctx.getDXRAMServiceAccessor().getService(ChunkService.class);
-        NameserviceService nameserviceService = p_ctx.getDXRAMServiceAccessor().getService(NameserviceService.class);
-        int numberOfWorkerNodes = p_ctx.getCtxData().getSlaveNodeIds().length;
 
-        chunkService.remove().remove(nameserviceService.getChunkID("GT", 100));
+        if (p_ctx.getCtxData().getSlaveId() == 0){
+            ChunkService chunkService = p_ctx.getDXRAMServiceAccessor().getService(ChunkService.class);
+            NameserviceService nameserviceService = p_ctx.getDXRAMServiceAccessor().getService(NameserviceService.class);
+            int numberOfWorkerNodes = p_ctx.getCtxData().getSlaveNodeIds().length;
 
-        for (int i=0; i<numberOfWorkerNodes; i++){
-            chunkService.remove().remove(nameserviceService.getChunkID("SID"+i, 100));
-            chunkService.remove().remove(nameserviceService.getChunkID("RC"+i, 100));
+            chunkService.remove().remove(nameserviceService.getChunkID("GT", 100));
+            chunkService.remove().remove(nameserviceService.getChunkID("WO", 100));
+
+            for (int i=0; i<numberOfWorkerNodes; i++){
+                chunkService.remove().remove(nameserviceService.getChunkID("SID"+i, 100));
+                chunkService.remove().remove(nameserviceService.getChunkID("RC"+i, 100));
+            }
         }
-
-
-
-
-
         return 0;
     }
